change frequency to GHz

+++ include/modules/network.hpp
@@ -73,7 +73,7 @@ class Network : public ALabel {
   int         cidr_;
   int32_t     signal_strength_dbm_;
   uint8_t     signal_strength_;
-  uint32_t    frequency_;
+  float       frequency_;
   uint32_t    route_priority;
 
   util::SleeperThread thread_;
